+++ scenes/input_output.gd
@@ -5,9 +5,16 @@ class_name InputOutput
 
 var id: int: set = _set_id, get = _get_id
 var value: float: set = _set_value, get = _get_value
+var desired_output: float: set = _set_desired_output, get = _get_desired_output
+var activation_function: String: get = _get_activation_function
+
+func _ready() -> void:
+	var layer_type = self.get_meta("layer_type", Global.LayerType.NeuronsInput)
+	if layer_type == Global.LayerType.NeuronsOutput:
+		%DesiredOutput.visible = true
 
 func _set_id(value: int) -> void:
-	self.title = Global.get_input_output_name(id + 1)
+	self.title = Global.get_input_output_name(value + 1)
 	id = value
 
 func _get_id() -> int:
@@ -16,10 +23,24 @@ func _get_id() -> int:
 func _set_value(value: float) -> void:
 	if !self.is_node_ready():
 		await self.ready
-	%Value.value = str(value)
+	if activation_function == "Same":
+		%Value.value = str(value)
+	elif activation_function == "Out>0":
+		%Value.value = str(1.0 if value > 0.0 else 0.0)
+	elif activation_function == "Out>0.5":
+		%Value.value = str(1.0 if value > 0.5 else 0.0)
 
 func _get_value() -> float:
 	return float(%Value.value)
+
+func _set_desired_output(value: float) -> void:
+	%DesiredOutput.value = str(value)
+
+func _get_desired_output() -> float:
+	return float(%DesiredOutput.value)
+
+func _get_activation_function() -> String:
+	return %ActivationFunction.value
 
 func update_value(value: float) -> void:
 	self.value = value