+++ scenes/neuron.gd
@@ -6,11 +6,15 @@ class_name Neuron
 var id: int: set = _set_id, get = _get_id
 var net: float: set = _set_net, get = _get_net
 var value: float: set = _set_value, get = _get_value
+var delta: float: set = _set_delta, get = _get_delta
 var activation_function: String: set = _set_activation_function, get = _get_activation_function
 var layer_node: Layer = null
 var in_neurons: Dictionary = {}
+var out_neurons: Dictionary = {}
 var inputs: Dictionary = {}
+var output: InputOutput = null
 var weights: Dictionary = {}
+var weights_nodes: Dictionary = {}
 
 signal value_updated(value: int)
 
@@ -21,15 +25,18 @@ func _append_weight(id: String, value: int) -> void:
 	input.initial_value = "1"
 	input.set_meta("id", id)
 	input.editable = false
+	weights_nodes[id] = input
 	%Weights.add_child(input)
 
 func _remove_weight(id: String) -> void:
 	weights.erase(id)
-	for i in range(%Weights.get_child_count()):
-		var node = %Weights.get_child(i)
-		if node.get_meta("id") == id:
-			node.free()
-			break
+	weights_nodes[id].free()
+
+func update_weight_on_input(id: String, value: float) -> void:
+	print(value)
+	print(str(value))
+	print("---------")
+	weights_nodes[id].value = str(value)
 
 func _set_id(value: int) -> void:
 	self.title = Global.get_neuron_name(value + 1)
@@ -51,6 +58,12 @@ func _set_value(value: float) -> void:
 func _get_value() -> float:
 	return float(%Output.value)
 
+func _set_delta(value: float) -> void:
+	%Delta.value = str(value)
+
+func _get_delta() -> float:
+	return float(%Delta.value)
+
 func _set_activation_function(value: String) -> void:
 	%ActivationFunction.value = value
 
@@ -67,37 +80,43 @@ func _on_show_button_up():
 	%Details.visible = false
 	%ValuesContainer.visible = true
 
-func connect_neuron(id: String, neuron: Neuron) -> bool:
+func connect_in_neuron(id: String, neuron: Neuron) -> bool:
 	if len(inputs) > 0:
 		return false
 	in_neurons[id] = neuron
 	_append_weight(id, 1.0)
 	return true
 
-func connect_input(id: String, input: InputOutput) -> bool:
+func connect_out_neuron(id: String, neuron: Neuron) -> bool:
+	out_neurons[id] = neuron
+	return true
+
+func connect_input(id: String, inputOutput: InputOutput) -> bool:
 	if len(in_neurons) > 0:
 		return false
-	inputs[id] = input
+	inputs[id] = inputOutput
 	_append_weight(id, 1.0)
 	return true
 
-func connect_output(input: InputOutput) -> bool:
-	value_updated.connect(input.update_value)
+func connect_output(inputOutput: InputOutput) -> bool:
+	output = inputOutput
+	value_updated.connect(inputOutput.update_value)
 	value_updated.emit(value)
 	return true
 
-func disconnect_neuron(id: String, neuron: Neuron) -> bool:
+func disconnect_in_neuron(id: String, neuron: Neuron) -> bool:
 	in_neurons.erase(id)
 	_remove_weight(id)
 	return true
 
-func disconnect_input(id: String, input: InputOutput) -> bool:
+func disconnect_input(id: String, inputOutput: InputOutput) -> bool:
 	inputs.erase(id)
 	_remove_weight(id)
 	return true
 
-func disconnect_output(input: InputOutput) -> bool:
-	value_updated.disconnect(input.update_value)
+func disconnect_output(inputOutput: InputOutput) -> bool:
+	output = null
+	value_updated.disconnect(inputOutput.update_value)
 	return true
 
 func _on_output_text_changed(inputControl: InputControl):
@@ -114,3 +133,31 @@ func feed_forward() -> void:
 		new_value += inputs_or_neurons[key].value * weights[key]
 	net = new_value
 	value = Global.calculate_neuron_output(net, activation_function)
+
+func calculate_delta() -> void:
+	var y_derivative = Global.calculate_derivative_neuron_output(net, activation_function)
+	if layer_node.final_layer:
+		delta = y_derivative * (output.desired_output - value)
+		return
+	var sigma = 0
+	for key in out_neurons:
+		sigma += out_neurons[key].weights[key] * out_neurons[key].delta
+	delta = y_derivative * sigma
+
+func feed_backward() -> void:
+	if id == 0:
+		for key in weights:
+			var delta_w = layer_node.manager.learning_rate * delta * inputs[key].value
+			var new_weight = weights[key] + delta_w
+			weights[key] = new_weight
+			update_weight_on_input(key, new_weight)
+		if output != null:
+			return
+	elif output != null:
+		return
+
+	for key in out_neurons:
+		var delta_w = layer_node.manager.learning_rate * out_neurons[key].delta * value
+		var new_weight = out_neurons[key].weights[key] + delta_w
+		out_neurons[key].weights[key] = new_weight
+		out_neurons[key].update_weight_on_input(id, new_weight)